Fix Jenkinsfile molecule hook call
The molecule hook itself, used within the user-defined preScenarioHook,
was triggering the error. I've updated the Jenkinsfile, as well as the
usage examples in the readme, to reflect this change.

Also, rhsm is slooooooooooooooooooooooooow so I switched it to
molecule_openstack_ci, and I added a sane default scenario array
for `molecule_scenarios`.